:recycle: Separate web surfing animation from state
